Spike how the server/client could propagate custom exceptions
This would make for a friendlier programming model than just
catching the RPC exceptions and dealing with exception codes
(at least when calling from .NET).

The hack to add this to StreamJsonRpc shows that this can be
added on top of the existing stuff (although code is throwaway
of course, it's a spike!), and it shouldn't break non-.NET
clients.

Related to https://github.com/microsoft/vs-streamjsonrpc/issues/468