Serialize and forward events across AppDomain
The `EventListener` can only receive events written in the
current `AppDomain`, so this showcases a way to forward them
across an `AppDomain` by serializing the entire event args.

The calling side (consumer, i.e. client app/output window)
deserializes the event args into a custom structure since
the `EventWrittenEventArgs` wasn't intended to be serialized,
but we can easily reconstruct the entire data, to the point
where the previous code barely changed to write the event
as a string message.

A "local" forwarder is also available just to showcase
how we could have a single approach for both.