Further showcase how to forward to DiagnosticSource
Rather than rendering directly from the received event payload,
convert it back to a form of `DiagnosticSource.Write` as it
happened on the original remote `AppDomain`, so that local
processing can also treat it uniformly with remote ones.

This wouldn't work for EventSource events though, so maybe it's
not really a useful thing. However, turning `DiagnosticSource.Write`
and `EventSource.Write(...)` into a uniform message string
would not be simple, since `DiagnosticSource` has no concept of
a message at all.